Refactor ItemMerge methods into mix-in module P::ItemMerge::Function
