+++ lib/poloxy.rb
@@ -28,6 +28,7 @@ require_relative 'poloxy/graph'
 require_relative 'poloxy/item'
 require_relative 'poloxy/item_merge'
 require_relative 'poloxy/item_merge/base'
+require_relative 'poloxy/item_merge/function'
 require_relative 'poloxy/logging'
 require_relative 'poloxy/message_container'
 require_relative 'poloxy/view_model/item'
